WIP: Snackbar refacto
