+++ packages/core/src/dialog-manager/models/dialog-config.ts
@@ -1,4 +1,4 @@
-import { RelativePosition } from "../../common";
+import { RelativePosition } from "../../common/common.models";
 
 export interface DialogConfig {
   /**
