+++ packages/core/src/snackbar-manager/models/snackbar-config.ts
@@ -2,7 +2,7 @@ import {
   HorizontalAlignement,
   RelativePosition,
   VerticalAlignement,
-} from "../../common";
+} from "../../common/common.models";
 
 export type SnackbarConfig = {
   /**
